+++ packages/vite-runtime-error-overlay/README.md
@@ -1,0 +1,15 @@
+# vite-runtime-error-overlay
+
+Vite plugin to show client runtime error via vite error overlay.
+Based on the idea from https://github.com/vitejs/vite/pull/6274#issuecomment-1087749460
+
+## demo
+
+TODO
+
+## development
+
+```sh
+pnpm build
+pnpm release
+```
