+++ packages/vite-runtime-error-overlay/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@hiogawa/vite-runtime-error-overlay",
+  "version": "0.0.0-pre.0",
+  "homepage": "https://github.com/hi-ogawa/unocss-preset-antd/tree/main/packages/vite-runtime-error-overlay",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/hi-ogawa/unocss-preset-antd/",
+    "directory": "packages/vite-runtime-error-overlay"
+  },
+  "license": "MIT",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs",
+      "types": "./dist/index.d.ts"
+    }
+  },
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsup",
+    "release": "pnpm publish --no-git-checks --access public"
+  },
+  "devDependencies": {
+    "vite": "^4.4.9"
+  },
+  "peerDependencies": {
+    "vite": "*"
+  }
+}
